AActualizando métodos

+++ raicesmx/src/main/java/org/generation/raicesmx/service/ArtesanoService.java
@@ -1,9 +1,9 @@
 package org.generation.raicesmx.service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
+import org.generation.raicesmx.exception.UserNotFoundException;
 import org.generation.raicesmx.model.ArtesanoEntity;
 import org.generation.raicesmx.repository.ArtesanoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,7 +26,13 @@ public class ArtesanoService {
 
 	// Mostrar un registro por id
 	public ArtesanoEntity getArtesano(Long id_artesano){
-	    return artesanoRepository.findById(id_artesano).orElse(null);
+	    return this.artesanoRepository.findById(id_artesano)
+	    		.orElseThrow(() -> new UserNotFoundException(id_artesano));
+	}
+	
+	// Mostrar por correo
+	public ArtesanoEntity findByEmail(String correo) {
+		return this.artesanoRepository.findByEmail(correo);
 	}
 	
 	// Crear un nuevo registro
@@ -36,16 +42,20 @@ public class ArtesanoService {
     
     // Eliminar un registro por id
     public void deleteArtesano(Long id) {
-        artesanoRepository.deleteById(id);
+    	if(this.artesanoRepository.existsById(id)) {
+    		artesanoRepository.deleteById(id);
+    	}else {
+    		throw new UserNotFoundException(id);
+    	}
+        
     }
     
     // Editar un registro
-    public ArtesanoEntity updateArtesano(ArtesanoEntity artesanoEntity) {
+    public ArtesanoEntity updateArtesano(ArtesanoEntity artesanoEntity, Long id) {
         Optional < ArtesanoEntity > artesanoDb = this.artesanoRepository.findById(artesanoEntity.getId_artesano());
 
         if (artesanoDb.isPresent()) {
             ArtesanoEntity artesanoUpdate = artesanoDb.get();
-            artesanoUpdate.setId_artesano(artesanoEntity.getId_artesano());
             artesanoUpdate.setNombre(artesanoEntity.getNombre());
             artesanoUpdate.setApellido( artesanoEntity.getApellido() );
             artesanoUpdate.setEmpresa( artesanoEntity.getEmpresa() );
@@ -55,11 +65,10 @@ public class ArtesanoService {
             artesanoUpdate.setTelefono( artesanoEntity.getTelefono() );
             artesanoUpdate.setCodigo_postal( artesanoEntity.getCodigo_postal() );
             artesanoUpdate.setEstado( artesanoEntity.getEstado() );
-            artesanoUpdate.setTipo_usuario( artesanoEntity.getTipo_usuario() );
             artesanoRepository.save(artesanoUpdate);
             return artesanoUpdate;
         } else {
-            throw new NoSuchElementException("No se encontró registro con id : " + artesanoEntity.getId_artesano());
+        	throw new UserNotFoundException(id);
         }
     }
