+++ raicesmx/src/main/java/org/generation/raicesmx/service/PedidoService.java
@@ -1,9 +1,10 @@
 package org.generation.raicesmx.service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
+import org.generation.raicesmx.exception.PedidoNotFoundException;
+import org.generation.raicesmx.exception.UserNotFoundException;
 import org.generation.raicesmx.model.PedidoEntity;
 import org.generation.raicesmx.repository.PedidoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,35 +24,40 @@ public class PedidoService {
         return this.pedidoRepository.findAll();
     }
 	
+	// Mostrar un registro por id
 	public PedidoEntity getPedido(Long id_pedido){
-	    return pedidoRepository.findById(id_pedido).orElse(null);
+	    return pedidoRepository.findById(id_pedido)
+	    		.orElseThrow(() -> new UserNotFoundException(id_pedido));
 	}
 
+	
     public PedidoEntity createPedido (PedidoEntity newPedido) {
         return this.pedidoRepository.save(newPedido);
     }
     
+    // Eliminar un registro por id
     public void deletePedido(Long id) {
-        pedidoRepository.deleteById(id);
+    	if(this.pedidoRepository.existsById(id)) {
+    		pedidoRepository.deleteById(id);
+    	}else {
+    		throw new PedidoNotFoundException(id);
+    	}
     }
     
     // Editar un pedido
-    public PedidoEntity updatePedido(PedidoEntity pedidoEntity) {
+    public PedidoEntity updatePedido(PedidoEntity pedidoEntity, Long id) {
         Optional < PedidoEntity > pedidoDb = this.pedidoRepository.findById(pedidoEntity.getId_pedido());
 
         if (pedidoDb.isPresent()) {
             PedidoEntity pedidoUpdate = pedidoDb.get();
-            pedidoUpdate.setId_pedido(pedidoEntity.getId_pedido());
             pedidoUpdate.setTotal(pedidoEntity.getTotal());
             pedidoUpdate.setDescripcion(pedidoEntity.getDescripcion());
             pedidoUpdate.setEstado_pedido(pedidoEntity.getEstado_pedido());
             pedidoUpdate.setFecha(pedidoEntity.getFecha());
-            pedidoUpdate.setClientes_idClientes(pedidoEntity.getClientes_idClientes());
-            pedidoUpdate.setProducto_idProducto(pedidoEntity.getProducto_idProducto());
             pedidoRepository.save(pedidoUpdate);
             return pedidoUpdate;
         } else {
-            throw new NoSuchElementException("No se encontró registro con id: " + pedidoEntity.getId_pedido());
+        	throw new PedidoNotFoundException(id);
         }
     }
 }
