+++ raicesmx/src/main/java/org/generation/raicesmx/service/ProductoService.java
@@ -1,9 +1,9 @@
 package org.generation.raicesmx.service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
+import org.generation.raicesmx.exception.ProductoNotFoundException;
 import org.generation.raicesmx.model.ProductoEntity;
 import org.generation.raicesmx.repository.ProductoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,36 +24,42 @@ public class ProductoService {
 	}
 
 	public ProductoEntity getProducto(Long id_producto){
-	    return productoRepository.findById(id_producto).orElse(null);
+	    return productoRepository.findById(id_producto)
+	    		.orElseThrow(() -> new ProductoNotFoundException(id_producto));
 	}
 
 	public ProductoEntity createProducto (ProductoEntity newProducto) {
 	    return this.productoRepository.save(newProducto);
 	}
     
+	// Eliminar un registro por id
     public void deleteProducto(Long id) {
-        productoRepository.deleteById(id);
+    	if(this.productoRepository.existsById(id)) {
+    		productoRepository.deleteById(id);
+    	}else {
+    		throw new ProductoNotFoundException(id);
+    	}
+        
     }
     
  // Editar un producto
-    public ProductoEntity updateProducto(ProductoEntity ProductoEntity) {
+    public ProductoEntity updateProducto(ProductoEntity ProductoEntity, Long id) {
         Optional < ProductoEntity > productoDb = this.productoRepository.findById(ProductoEntity.getId_producto());
 
         if (productoDb.isPresent()) {
             ProductoEntity productoUpdate = productoDb.get();
-            productoUpdate.setId_producto(ProductoEntity.getId_producto());
             productoUpdate.setImagen_url(ProductoEntity.getImagen_url());
             productoUpdate.setNombre(ProductoEntity.getNombre());
             productoUpdate.setPrecio(ProductoEntity.getPrecio());
             productoUpdate.setStock(ProductoEntity.getStock());
-            productoUpdate.setCategorias_idCategorias(ProductoEntity.getCategorias_idCategorias());
+           
             productoUpdate.setStatus(ProductoEntity.getStatus());
             productoUpdate.setDescripcion(ProductoEntity.getDescripcion());
-            productoUpdate.setArtesano_idArtesano(ProductoEntity.getArtesano_idArtesano());
+            
             productoRepository.save(productoUpdate);
             return productoUpdate;
         } else {
-            throw new NoSuchElementException("No se encontró registro con id: " + ProductoEntity.getId_producto());
+        	throw new ProductoNotFoundException(id);
         }
     }
 }
